+++ waggle/waggle_native.h
@@ -9,7 +9,7 @@ public:
     PrintfWriter(const char *fmt) : fmt(fmt) {
     }
 
-    int Write(const byte *data, int size) {
+    size_t Write(const byte *data, size_t size) {
         for (int i = 0; i < size; i++) {
             printf(fmt, (byte)(data[i]));
         }
